+++ src/main/java/io/github/seanwu1105/umldiagrameditor/diagram/object/CompositeObject.java
@@ -15,8 +15,8 @@ public class CompositeObject implements UmlObject {
 
     @Override
     public @NotNull Position getPosition() {
-        var xMin = Integer.MAX_VALUE;
-        var yMin = Integer.MAX_VALUE;
+        var xMin = Double.MAX_VALUE;
+        var yMin = Double.MAX_VALUE;
         for (final var child : getChildren()) {
             xMin = Math.min(xMin, child.getPosition().getX());
             yMin = Math.min(yMin, child.getPosition().getY());
@@ -25,8 +25,8 @@ public class CompositeObject implements UmlObject {
     }
 
     @Override
-    public int getWidth() {
-        var xMax = Integer.MIN_VALUE;
+    public double getWidth() {
+        var xMax = Double.MIN_VALUE;
         for (final var child : getChildren()) {
             xMax = Math.max(xMax, child.getPosition().getX() + child.getWidth());
         }
@@ -34,8 +34,8 @@ public class CompositeObject implements UmlObject {
     }
 
     @Override
-    public int getHeight() {
-        var yMax = Integer.MIN_VALUE;
+    public double getHeight() {
+        var yMax = Double.MIN_VALUE;
         for (final var child : getChildren()) {
             yMax = Math.max(yMax, child.getPosition().getY() + child.getHeight());
         }
