+++ src/main/java/io/github/seanwu1105/umldiagrameditor/diagram/object/UmlObject.java
@@ -8,9 +8,9 @@ public interface UmlObject {
 
     @NotNull Position getPosition();
 
-    int getHeight();
+    double getHeight();
 
-    int getWidth();
+    double getWidth();
 
     @Nullable
     CompositeObject getParent();
